requirements.txt
# Web framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1

# Scheduling
APScheduler==3.10.4

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
lxml==4.9.3

# AI
google-generativeai==0.3.2

# Configuration
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.0
